requirements.txt
flask
transformers
torch
pillow
requests
timm